pipecat-quickstart/pyproject.toml
[project]
name = "pipecat-quickstart"
version = "0.1.0"
description = "Quickstart example for building voice AI bots with Pipecat"
requires-python = ">=3.10"
dependencies = [
    "pipecat-ai[webrtc,daily,silero,deepgram,google,cartesia,tavus,local-smart-turn-v3,runner]",
    "pipecat-ai-cli",
    "supabase",
    "websockets",
    "google.generativeai"
]

[dependency-groups]
dev = [
    "pyright>=1.1.404,<2",
    "ruff>=0.12.11,<1",
]

[tool.ruff]
line-length = 100
[tool.ruff.lint]
select = ["I"]